refactor: Split search_author
