+++ src/command/search.rs
@@ -1,0 +1,1 @@
+pub mod author;
